Revert "Move typechain output to dist/typechain/ and fix imports"
This reverts commit 4eaaa779fbff1f86163255f9ca2d0291f9e6061d.
