+++ hardhat/test/00-ape-protocol/02-ape-router.ts
@@ -13,7 +13,7 @@ import {
   ApeVaultWrapperImplementation,
   ERC20,
   VaultAPI,
-} from '../../dist/typechain';
+} from '../../typechain';
 import { Account } from '../utils/account';
 import { createApeVault } from '../utils/ApeVault/createApeVault';
 import { DeploymentInfo, deployProtocolFixture } from '../utils/deployment';
